Fix remove_dir_all

+++ src/yt_dlp.rs
@@ -140,7 +140,9 @@ impl YtDlp {
 
 		log::info!("Downloading yt-dlp release {}", tag_name);
 
-		tokio::fs::remove_dir_all("yt_dlp_exe").await?;
+		if Path::new("yt_dlp_exe").is_dir() {
+			tokio::fs::remove_dir_all("yt_dlp_exe").await?;
+		}
 
 		tokio::fs::create_dir_all("yt_dlp_exe").await?;
